+++ src/__tests__/FormGroup.spec.ts
@@ -1,7 +1,7 @@
 import { assert as t } from "chai";
 import { useStrict } from "mobx";
 import Field from "../Field";
-import Form from "../Form";
+import FormGroup from "../FormGroup";
 
 useStrict(true);
 
@@ -9,7 +9,7 @@ describe("Form", () => {
   it("should initialize via constructor", () => {
     const fa = new Field("a");
     const fb = new Field("b");
-    const form = new Form([fa, fb]);
+    const form = new FormGroup([fa, fb]);
 
     t.equal(form.valid, true);
   });
@@ -17,7 +17,7 @@ describe("Form", () => {
   it("should return valid if fields are valid", () => {
     const fa = new Field("a");
     const fb = new Field("b");
-    const form = new Form();
+    const form = new FormGroup();
     form.addFields(fa, fb);
 
     t.equal(form.valid, true);
@@ -26,7 +26,7 @@ describe("Form", () => {
   it("should return invalid if one field is invalid", () => {
     const fa = new Field("a");
     const fb = new Field("b", { required: true });
-    const form = new Form();
+    const form = new FormGroup();
     form.addFields(fa, fb);
 
     fb.setValue("yes");
@@ -38,7 +38,7 @@ describe("Form", () => {
   it("should add fields", () => {
     const fa = new Field("a");
     const fb = new Field("b");
-    const form = new Form();
+    const form = new FormGroup();
     form.addFields(fa, fb);
 
     t.equal(Object.keys(form.fields).length > 0, true);
@@ -47,7 +47,7 @@ describe("Form", () => {
   it("should getField by name", () => {
     const fa = new Field("foo");
     const fb = new Field("bar");
-    const form = new Form();
+    const form = new FormGroup();
     form.addFields(fa, fb);
 
     t.equal(form.fields["foo"].name, "foo");
