+++ src/index.ts
@@ -1,3 +1,3 @@
 export { default as Field } from "./Field";
-export { default as Form } from "./Form";
+export { default as Form } from "./FormGroup";
 export { default as SimpleForm } from "./SimpleForm";
